+++ packages/core/src/components/BrandConcierge/index.js
@@ -30,8 +30,8 @@ const createConciergeComponent = ({
   lifecycle,
   cookieTransfer,
   createResponse,
+  platformServices,
 }) => {
-  const { fetch } = window;
   const session = {
     id: getConciergeSessionCookie({ loggingCookieJar, config }),
   };
@@ -63,13 +63,16 @@ const createConciergeComponent = ({
     sendConversationServiceRequest,
     decodeKndctrCookie,
     session,
+    getPageLocation: platformServices.globals.getPageLocation,
   });
 
   return {
     lifecycle: {
       onBeforeEvent({ event }) {
         if (config.conversation.collectSources) {
-          const parsedParams = queryString.parse(window.location.search);
+          const parsedParams = queryString.parse(
+            platformServices.globals.getLocationSearch(),
+          );
           const source = parsedParams[SOURCES_QUERY_PARAM];
           if (source) {
             event.mergeXdm({ channel: { referringSource: source } });