+++ packages/core/test/unit/specs/components/Consent/createConsentHashStore.spec.js
@@ -16,17 +16,13 @@ import createConsentHashStore from "../../../../../src/components/Consent/create
 const CONSENT_IN = {
   standard: "Adobe",
   version: "1.0",
-  value: {
-    general: "in",
-  },
+  value: { general: "in" },
 };
 
 const CONSENT_OUT = {
   standard: "Adobe",
   version: "1.0",
-  value: {
-    general: "out",
-  },
+  value: { general: "out" },
 };
 
 describe("createConsentHashStore", () => {
@@ -34,13 +30,11 @@ describe("createConsentHashStore", () => {
   let subject;
   beforeEach(() => {
     storage = {
-      getItem: vi.fn(),
-      setItem: vi.fn(),
-      clear: vi.fn(),
+      getItem: vi.fn().mockResolvedValue(null),
+      setItem: vi.fn().mockResolvedValue(true),
+      clear: vi.fn().mockResolvedValue(true),
     };
-    subject = createConsentHashStore({
-      storage,
-    });
+    subject = createConsentHashStore({ storage });
   });
 
   it("clears", () => {
@@ -48,29 +42,29 @@ describe("createConsentHashStore", () => {
     expect(storage.clear).toHaveBeenCalled();
   });
 
-  it("is new when storage is empty", () => {
-    storage.getItem.mockReturnValue(null);
+  it("is new when storage is empty", async () => {
+    storage.getItem.mockResolvedValue(null);
     const consentHashes = subject.lookup([CONSENT_IN]);
-    expect(consentHashes.isNew()).toBe(true);
+    expect(await consentHashes.isNew()).toBe(true);
   });
 
-  it("saves the hash", () => {
+  it("saves the hash", async () => {
     const consentHashes = subject.lookup([CONSENT_IN]);
-    consentHashes.save();
+    await consentHashes.save();
     expect(storage.setItem).toHaveBeenCalledWith("Adobe.1.0", "e385c3ab");
   });
 
-  it("is not new when lookup is the same", () => {
-    storage.getItem.mockReturnValue("e385c3ab");
+  it("is not new when lookup is the same", async () => {
+    storage.getItem.mockResolvedValue("e385c3ab");
     const consentHashes = subject.lookup([CONSENT_IN]);
-    expect(consentHashes.isNew()).toBe(false);
+    expect(await consentHashes.isNew()).toBe(false);
     expect(storage.getItem).toHaveBeenCalledWith("Adobe.1.0");
   });
 
-  it("is new when lookup is different", () => {
-    storage.getItem.mockReturnValue("e385c3ab");
+  it("is new when lookup is different", async () => {
+    storage.getItem.mockResolvedValue("e385c3ab");
     const consentHashes = subject.lookup([CONSENT_OUT]);
-    expect(consentHashes.isNew()).toBe(true);
+    expect(await consentHashes.isNew()).toBe(true);
     expect(storage.getItem).toHaveBeenCalledWith("Adobe.1.0");
   });
 });